requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pulsar==1.4.1
PyYAML==3.12
Werkzeug==0.11.11
